+++ modules/spring-web/src/main/kotlin/com/explyt/spring/web/SpringWebClasses.kt
@@ -28,6 +28,8 @@ object SpringWebClasses {
     const val RESPONSE_BODY = "org.springframework.web.bind.annotation.ResponseBody"
     const val REQUEST_MAPPING = "org.springframework.web.bind.annotation.RequestMapping"
     const val PATH_VARIABLE = "org.springframework.web.bind.annotation.PathVariable"
+    const val REQUEST_PARAM = "org.springframework.web.bind.annotation.RequestParam"
+    const val REQUEST_BODY = "org.springframework.web.bind.annotation.RequestBody"
     const val RESPONSE_ENTITY = "org.springframework.http.ResponseEntity"
     const val FEIGN_CLIENT = "org.springframework.cloud.openfeign.FeignClient"